requirements.txt
autobahn==22.7.1
pyOpenSSL==22.0.0
pyserde[toml]==0.9.2
service_identity==21.1.0
toml==0.10.2
Twisted==22.8.0